+++ src/components/Accordion/AccordionSection/types.ts
@@ -1,4 +1,6 @@
 export type AccordionSectionProps = {
   sectionId: string;
   className?: string;
+  disableSelection?: boolean;
+  disableExpansion?: boolean;
 };