+++ src/components/Popups/Alerts/components/AlertFooter.tsx
@@ -1,11 +1,13 @@
 import React, { useEffect } from "react";
 
+import { Icon } from "../../../Miscellaneous/Icon";
 import { AlertFooterProps } from "../types";
 
 export function AlertFooter<TData = undefined>({
   data,
   confirmBtnText = "Confirm",
   confirmBtnDisabled = false,
+  confirmBtnBusy = false,
   disableConfirmOnEnter = false,
   onConfirm,
   declineBtnText = "Close",
@@ -13,26 +15,26 @@ export function AlertFooter<TData = undefined>({
   onDecline,
 }: AlertFooterProps<TData>) {
   useEffect(() => {
-    if (disableConfirmOnEnter) {
-      return;
-    }
-
     const handleKeyDown = (event: KeyboardEvent) => {
-      if (onConfirm && event.code === "Enter") {
-        onConfirm(data as unknown as TData);
+      if (event.code === "Enter") {
+        if (confirmBtnBusy || confirmBtnDisabled || disableConfirmOnEnter) {
+          return;
+        }
+
+        if (onConfirm) {
+          event.preventDefault();
+          event.stopPropagation();
+          onConfirm(data as unknown as TData);
+        }
       }
     };
 
     document.addEventListener("keydown", handleKeyDown);
 
     return () => {
-      if (disableConfirmOnEnter) {
-        return;
-      }
-
       document.removeEventListener("keydown", handleKeyDown);
     };
-  }, []);
+  }, [disableConfirmOnEnter, onConfirm, data, confirmBtnBusy, confirmBtnDisabled]);
 
   if (!onConfirm && !onDecline) {
     return null;
@@ -42,10 +44,13 @@ export function AlertFooter<TData = undefined>({
     <div className="m-alert-footer">
       {onConfirm && (
         <button
-          disabled={confirmBtnDisabled}
+          disabled={confirmBtnDisabled || confirmBtnBusy}
           onClick={() => onConfirm(data as unknown as TData)}
           className="m-alert-confirm-button"
         >
+          {confirmBtnBusy && (
+            <Icon icon={["fas", "circle-notch"]} className="fa-spin m-alert-confirm-button-busy-icon" />
+          )}
           {confirmBtnText}
         </button>
       )}