+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### Release date 07.05.2026 - Release 2.1.7
+
+- [[#274]](https://github.com/Marcin-Migdal/m-component-library/issues/274) **[PATCH]** Add stopPropagation to DropdownMenu click to prevent external event capture
+- [[#272]](https://github.com/Marcin-Migdal/m-component-library/issues/272) **[PATCH]** Add disableSelection and disableExpansion props to Accordion component
+- [[#264]](https://github.com/Marcin-Migdal/m-component-library/issues/264) **[PATCH]** Add checked property to Form's controlled register change result
+- [[#265]](https://github.com/Marcin-Migdal/m-component-library/issues/265) **[PATCH]** Imagefield add src prop to pass image url that should be displayed
+- [[#266]](https://github.com/Marcin-Migdal/m-component-library/issues/266) **[PATCH]** Add busy icon to the confirm btn in alert footer
+
 ### Release date 15.02.2026 - Release 2.1.6
 
 - [[#250]](https://github.com/Marcin-Migdal/m-component-library/issues/250) **[PATCH]** Improve DropdownMenu to use DropdownSubMenu with template